--- a/go/tools/macrobench/sql.go
+++ b/go/tools/macrobench/sql.go
@@ -17,63 +17,121 @@ limitations under the License.
 package macrobench
 
 import (
+	"database/sql"
 	"errors"
 	"strings"
 
+	"github.com/vitessio/arewefastyet/go/exec/metrics"
 	"github.com/vitessio/arewefastyet/go/storage"
 	"github.com/vitessio/arewefastyet/go/storage/mysql"
 )
 
-// getResultsForGitRefAndPlanner returns a slice of details based on the given git ref
-// and macro benchmark Type.
-func getResultsForGitRefAndPlanner(macroType string, ref string, planner PlannerVersion, client storage.SQLClient) (macrodetails detailsArray, err error) {
+// getExecutionGroupResults the results of an execution group
+func getExecutionGroupResults(macroType string, ref string, planner PlannerVersion, client storage.SQLClient) (executionGroupResults, error) {
 	upperMacroType := strings.ToUpper(macroType)
-	query := "SELECT " +
-		"info.macrobenchmark_id, e.git_ref, e.source, e.finished_at, IFNULL(info.exec_uuid, ''), " +
-		"results.tps, results.latency, results.errors, results.reconnects, results.time, results.threads, " +
-		"results.total_qps, results.reads_qps, results.writes_qps, results.other_qps " +
-		"FROM execution AS e, macrobenchmark AS info, macrobenchmark_results AS results " +
-		"WHERE e.uuid = info.exec_uuid " +
-		"AND e.status = \"finished\" " +
-		"AND e.git_ref = ? " +
-		"AND info.vtgate_planner_version = ? " +
-		"AND info.macrobenchmark_id = results.macrobenchmark_id " +
-		"AND info.type = ?"
+	query := `
+        SELECT 
+            IFNULL(e.uuid, '') AS exec_uuid, 
+            results.tps, 
+            results.latency, 
+            results.errors, 
+            results.reconnects, 
+            results.time, 
+            results.threads, 
+            results.total_qps, 
+            results.reads_qps, 
+            results.writes_qps, 
+            results.other_qps, 
+            m.name AS metric_name, 
+            m.value AS metric_value
+        FROM 
+            execution AS e
+        JOIN 
+            macrobenchmark AS info ON e.uuid = info.exec_uuid
+        JOIN 
+            macrobenchmark_results AS results ON info.macrobenchmark_id = results.macrobenchmark_id
+        LEFT JOIN 
+            metrics AS m ON e.uuid = m.exec_uuid
+        WHERE 
+            e.status = 'finished'
+            AND e.git_ref = ? 
+            AND info.vtgate_planner_version = ? 
+            AND info.type = ?
+        ORDER BY 
+            e.uuid, m.name
+    `
 
-	result, err := client.Read(query, ref, planner, upperMacroType)
+	rows, err := client.Read(query, ref, planner, upperMacroType)
 	if err != nil {
-		return nil, err
+		return executionGroupResults{}, err
 	}
-	defer result.Close()
-	for result.Next() {
-		var res details
-		err = result.Scan(
-			&res.ID,
-			&res.GitRef,
-			&res.Source,
-			&res.CreatedAt,
-			&res.ExecUUID,
-			&res.Result.TPS,
-			&res.Result.Latency,
-			&res.Result.Errors,
-			&res.Result.Reconnects,
-			&res.Result.Time,
-			&res.Result.Threads,
-			&res.Result.QPS.Total,
-			&res.Result.QPS.Reads,
-			&res.Result.QPS.Writes,
-			&res.Result.QPS.Other,
+	defer rows.Close()
+
+	results := executionGroupResults{GitRef: ref}
+	var execRes *executionResults
+	var currentExecUUID string
+
+	for rows.Next() {
+		var (
+			execUUID    string
+			sr          sysbenchResult
+			metricName  sql.NullString
+			metricValue sql.NullFloat64
+		)
+
+		err := rows.Scan(
+			&execUUID, &sr.TPS, &sr.Latency, &sr.Errors, &sr.Reconnects, &sr.Time, &sr.Threads, &sr.QPS.Total,
+			&sr.QPS.Reads, &sr.QPS.Writes, &sr.QPS.Other, &metricName, &metricValue,
 		)
 		if err != nil {
-			return nil, err
+			return executionGroupResults{}, err
 		}
-		macrodetails = append(macrodetails, res)
+
+		// If execUUID is different it means we are looking at another set of results
+		// we then add the current execRes to our results, and start again with a new executionResults
+		if currentExecUUID != execUUID {
+			if execRes != nil {
+				results.Results = append(results.Results, execRes.Result)
+				results.Metrics = append(results.Metrics, execRes.Metrics)
+			}
+			execRes = &executionResults{
+				Result: sr,
+				Metrics: metrics.ExecutionMetrics{
+					ComponentsCPUTime:            make(map[string]float64),
+					ComponentsMemStatsAllocBytes: make(map[string]float64),
+				},
+			}
+			currentExecUUID = execUUID
+		}
+
+		// For each execution we will have multiple rows since we are doing a LEFT JOIN on metrics
+		// here we just all the metrics value to the executionResults, later when we are done consuming
+		// all the metrics for our current execUUID we will create a new executionResults
+		if metricName.Valid {
+			switch {
+			case metricName.String == "TotalComponentsCPUTime":
+				execRes.Metrics.TotalComponentsCPUTime = metricValue.Float64
+			case metricName.String == "TotalComponentsMemStatsAllocBytes":
+				execRes.Metrics.TotalComponentsMemStatsAllocBytes = metricValue.Float64
+			case strings.HasPrefix(metricName.String, "ComponentsCPUTime."):
+				execRes.Metrics.ComponentsCPUTime[strings.Split(metricName.String, ".")[1]] = metricValue.Float64
+			case strings.HasPrefix(metricName.String, "ComponentsMemStatsAllocBytes."):
+				execRes.Metrics.ComponentsMemStatsAllocBytes[strings.Split(metricName.String, ".")[1]] = metricValue.Float64
+			}
+		}
 	}
-	return macrodetails, nil
+
+	// This is used to add the results of our very last execUUID since we will exit the
+	// previous loop before adding the results.
+	if execRes != nil {
+		results.Results = append(results.Results, execRes.Result)
+		results.Metrics = append(results.Metrics, execRes.Metrics)
+	}
+	return results, nil
 }
 
-func getResultsLastXDays(macroType string, source string, planner PlannerVersion, lastDays int, client storage.SQLClient) (macrodetails detailsArray, err error) {
-	macrodetails = []details{}
+func getResultsLastXDays(macroType string, source string, planner PlannerVersion, lastDays int, client storage.SQLClient) (macrodetails executionResultsArray, err error) {
+	macrodetails = []executionResults{}
 	upperMacroType := strings.ToUpper(macroType)
 	query := "SELECT " +
 		"info.macrobenchmark_id, e.git_ref, e.source, e.finished_at, IFNULL(e.uuid, ''), " +
@@ -95,7 +153,7 @@ func getResultsLastXDays(macroType string, source string, planner PlannerVersion
 	}
 	defer result.Close()
 	for result.Next() {
-		var res details
+		var res executionResults
 		err = result.Scan(
 			&res.ID,
 			&res.GitRef,
@@ -121,7 +179,7 @@ func getResultsLastXDays(macroType string, source string, planner PlannerVersion
 	return macrodetails, nil
 }
 
-func getSummaryLastXDays(macroType string, source string, planner PlannerVersion, lastDays int, client storage.SQLClient) (results detailsArray, err error) {
+func getSummaryLastXDays(macroType string, source string, planner PlannerVersion, lastDays int, client storage.SQLClient) (results executionResultsArray, err error) {
 	upperMacroType := strings.ToUpper(macroType)
 	query := "SELECT " +
 		"info.macrobenchmark_id, e.git_ref, results.total_qps, IFNULL(e.uuid, '') " +
@@ -142,7 +200,7 @@ func getSummaryLastXDays(macroType string, source string, planner PlannerVersion
 	}
 	defer result.Close()
 	for result.Next() {
-		var res details
+		var res executionResults
 		err = result.Scan(&res.ID, &res.GitRef, &res.Result.QPS.Total, &res.ExecUUID)
 		if err != nil {
 			return nil, err
@@ -155,13 +213,13 @@ func getSummaryLastXDays(macroType string, source string, planner PlannerVersion
 // insertToMySQL inserts the given MacroBenchmarkResult to MySQL using a *mysql.Client.
 // The MacroBenchmarkResults gets added in one of macrobenchmark's children tables.
 // Depending on the MacroBenchmarkType, the insert will be routed to a specific children table.
-// The children table qps is also inserted.
-func (mbr *result) insertToMySQL(macrobenchmarkID int, client storage.SQLClient) error {
+// The children table sysbenchQPS is also inserted.
+func (mbr *sysbenchResult) insertToMySQL(macrobenchmarkID int, client storage.SQLClient) error {
 	if client == nil {
 		return errors.New(mysql.ErrorClientConnectionNotInitialized)
 	}
 
-	// insert result
+	// insert sysbenchResult
 	queryResult := "INSERT INTO macrobenchmark_results(macrobenchmark_id, queries, tps, latency, errors, reconnects, time, threads, total_qps, reads_qps, writes_qps, other_qps) VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)"
 	_, err := client.Write(queryResult, macrobenchmarkID, mbr.Queries, mbr.TPS, mbr.Latency, mbr.Errors, mbr.Reconnects, mbr.Time, mbr.Threads, mbr.QPS.Total, mbr.QPS.Reads, mbr.QPS.Writes, mbr.QPS.Other)
 	if err != nil {