This PR revamps some of our endpoints that get and compare results. The performance were really poor, with for instance a 20 seconds response time for the compare benchmark page, and 25 seconds for the daily results. The response time of these two endpoints was reduced by 88% and 99% on average.

**Before**
```
[GIN] 2024/07/25 - 19:19:40 | 200 | 23.598892375s |             ::1 | GET      "/api/daily?workload=oltp"
[GIN] 2024/07/25 - 19:20:03 | 200 |   23.0457765s |             ::1 | GET      "/api/daily?workload=oltp"
[GIN] 2024/07/25 - 19:20:22 | 200 | 19.204524417s |             ::1 | GET      "/api/daily?workload=oltp"
[GIN] 2024/07/25 - 19:20:24 | 200 | 22.733138458s |             ::1 | GET      "/api/daily?workload=oltp"
[GIN] 2024/07/25 - 19:20:32 | 200 |  19.07924375s |             ::1 | GET      "/api/daily?workload=oltp"
[GIN] 2024/07/25 - 19:20:41 | 200 | 18.771850875s |             ::1 | GET      "/api/daily?workload=oltp"
[GIN] 2024/07/25 - 19:20:59 | 200 | 18.885992375s |             ::1 | GET      "/api/daily?workload=oltp"
```

**After**
```
[GIN] 2024/07/25 - 19:18:23 | 200 |  638.252792ms |             ::1 | GET      "/api/daily?workload=oltp"
[GIN] 2024/07/25 - 19:18:24 | 200 |  176.436084ms |             ::1 | GET      "/api/daily?workload=oltp"
[GIN] 2024/07/25 - 19:18:26 | 200 |  165.110584ms |             ::1 | GET      "/api/daily?workload=oltp"
[GIN] 2024/07/25 - 19:18:26 | 200 |  155.194375ms |             ::1 | GET      "/api/daily?workload=oltp"
[GIN] 2024/07/25 - 19:18:27 | 200 |  150.058625ms |             ::1 | GET      "/api/daily?workload=oltp"
[GIN] 2024/07/25 - 19:18:28 | 200 |    144.3535ms |             ::1 | GET      "/api/daily?workload=oltp"
[GIN] 2024/07/25 - 19:18:29 | 200 |  125.182625ms |             ::1 | GET      "/api/daily?workload=oltp"
[GIN] 2024/07/25 - 19:18:29 | 200 |  119.004083ms |             ::1 | GET      "/api/daily?workload=oltp"
[GIN] 2024/07/25 - 19:18:30 | 200 |  120.047583ms |             ::1 | GET      "/api/daily?workload=oltp"
[GIN] 2024/07/25 - 19:18:31 | 200 |     131.402ms |             ::1 | GET      "/api/daily?workload=oltp"
[GIN] 2024/07/25 - 19:18:32 | 200 |  119.589542ms |             ::1 | GET      "/api/daily?workload=oltp"
[GIN] 2024/07/25 - 19:18:33 | 200 |  142.467084ms |             ::1 | GET      "/api/daily?workload=oltp"
[GIN] 2024/07/25 - 19:18:33 | 200 |    138.0815ms |             ::1 | GET      "/api/daily?workload=oltp"
[GIN] 2024/07/25 - 19:18:34 | 200 |  131.591875ms |             ::1 | GET      "/api/daily?workload=oltp"
[GIN] 2024/07/25 - 19:18:35 | 200 |  134.980542ms |             ::1 | GET      "/api/daily?workload=oltp"
```